implemented purchase

+++ src/main/java/com/ilario/sparkmart/dto/PurchaseDTO.java
@@ -3,5 +3,5 @@ package com.ilario.sparkmart.dto;
 import java.util.Map;
 import java.util.UUID;
 
-public record PurchaseDTO(UUID userId, Map<UUID, Integer> products) {
+public record PurchaseDTO(UUID userId, Map<UUID, Short> products) {
 }
